+++ tests/Feature/Http/Controllers/Webhooks/PaymentGateways/LoadsRequestFixture.php
@@ -10,7 +10,10 @@ trait LoadsRequestFixture
      */
     public function getRequestData(string $fixtureFilename): array
     {
-        $fixtureFile = __DIR__.DIRECTORY_SEPARATOR.$fixtureFilename;
-        return json_decode(file_get_contents($fixtureFile), true);
+        $reflector = new \ReflectionClass(__CLASS__);
+        $dir = dirname($reflector->getFileName());
+
+        $fixtureFilepath = $dir.DIRECTORY_SEPARATOR.$fixtureFilename;
+        return json_decode(file_get_contents($fixtureFilepath), true);
     }
 }
