+++ tests/Feature/Http/Controllers/Webhooks/PaymentGateways/Subscriptions/created.subscription.json
@@ -1,0 +1,48 @@
+{
+  "id": "sbs_2abee7ad934beb93",
+  "state": "active",
+  "tracking_id": null,
+  "device_id": null,
+  "created_at": "2019-05-16T22:16:09.654Z",
+  "renew_at": "2019-06-16T22:16:30.474Z",
+  "active_to": "2019-06-16T22:16:30.474Z",
+  "card": {
+    "token": "6a0a063c-ddcb-45ca-8037-042cbac092df"
+  },
+  "customer": {
+    "id": "cst_388881961e365744"
+  },
+  "paid_billing_cycles": 1,
+  "number_failed_payment_attempts": 0,
+  "additional_data": {
+    "campaign_id": 1,
+    "donator_id": 3
+  },
+  "plan": {
+    "id": "pln_91ca7d31abf4b7d8",
+    "title": "25BYN for democracy",
+    "currency": "BYN",
+    "language": null,
+    "infinite": false,
+    "billing_cycles": 1,
+    "trial": {
+      "amount": null,
+      "interval": null,
+      "interval_unit": null
+    },
+    "plan": {
+      "amount": 2500,
+      "interval": 1,
+      "interval_unit": "month"
+    },
+    "number_payment_attempts": 3,
+    "test": true
+  },
+  "last_transaction": {
+    "uid": "13385020-7a35ad2dba",
+    "status": "successful",
+    "message": "Successfully processed",
+    "created_at": "2019-05-16T22:16:25.650Z"
+  },
+  "event": "created.subscription"
+}
